requirements.txt
numpy==1.16.4
opencv_python==4.1.0.25
six==1.11.0
Keras==2.2.4
Pillow==6.1.0
